% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{as.matrix.MXNDArray}
\alias{as.matrix.MXNDArray}
\title{as.matrix operator overload of mx.ndarray}
\usage{
\method{as.matrix}{MXNDArray}(nd)
}
\arguments{
\item{nd}{The mx.ndarray}
}
\description{
as.matrix operator overload of mx.ndarray
}

