% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{dim.MXNDArray}
\alias{dim.MXNDArray}
\title{Dimension operator overload of mx.ndarray}
\usage{
\method{dim}{MXNDArray}(nd)
}
\arguments{
\item{nd}{The mx.ndarray}
}
\description{
Dimension operator overload of mx.ndarray
}

