% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.graph.R
\name{graph.viz}
\alias{graph.viz}
\title{Convert symbol to dot object for visualization purpose.}
\usage{
graph.viz(model, graph.title = "Computation graph",
  graph.title.font.name = "Helvetica", graph.title.font.size = 30,
  graph.width.px = 500, graph.height.px = 500)
}
\arguments{
\item{model}{a \code{string} representing the path to a file containing the \code{JSon} of a model dump or the actual model dump.}

\item{graph.title}{a \code{string} displayed on top of the viz.}

\item{graph.title.font.name}{a \code{string} representing the font to use for the title.}

\item{graph.title.font.size}{a \code{numeric} representing the size of the font to use for the title.}

\item{graph.width.px}{a \code{numeric} representing the size (width) of the graph. In pixels}

\item{graph.height.px}{a \code{numeric} representing the size (height) of the graph. In pixels}
}
\value{
a graph object ready to be displayed with the \code{print} function.
}
\description{
Convert symbol to dot object for visualization purpose.
}

