% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{length.MXNDArray}
\alias{length.MXNDArray}
\title{Length operator overload of mx.ndarray}
\usage{
\method{length}{MXNDArray}(nd)
}
\arguments{
\item{nd}{The mx.ndarray}
}
\description{
Length operator overload of mx.ndarray
}

