% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback.R
\name{mx.callback.log.train.metric}
\alias{mx.callback.log.train.metric}
\title{Log training metric each period}
\usage{
mx.callback.log.train.metric(period)
}
\description{
Log training metric each period
}

