% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{mx.callback.save.checkpoint}
\alias{mx.callback.save.checkpoint}
\title{Save checkpoint to files each period iteration.}
\usage{
mx.callback.save.checkpoint(prefix, period = 1)
}
\arguments{
\item{prefix}{The prefix of the model checkpoint.}
}
\description{
Save checkpoint to files each period iteration.
}

