% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{mx.cpu}
\alias{mx.cpu}
\title{Create a mxnet CPU context.}
\arguments{
\item{dev.id}{optional, default=0
The device ID, this is meaningless for CPU, included for interface compatiblity.}
}
\value{
The CPU context.
}
\description{
Create a mxnet CPU context.
}

