% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executor.R
\name{mx.exec.forward}
\alias{mx.exec.forward}
\title{Peform an forward on the executors
This function will MUTATE the state of exec}
\usage{
mx.exec.forward(exec, is.train = TRUE)
}
\description{
Peform an forward on the executors
This function will MUTATE the state of exec
}

