% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executor.R
\name{mx.exec.update.aux.arrays}
\alias{mx.exec.update.aux.arrays}
\title{Update the executors with new arrays
This function will MUTATE the state of exec}
\usage{
mx.exec.update.aux.arrays(exec, arg.arrays, match.name = FALSE,
  skip.null = FALSE)
}
\description{
Update the executors with new arrays
This function will MUTATE the state of exec
}

