% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{mx.init.internal.default}
\alias{mx.init.internal.default}
\title{Internal default value initialization scheme.}
\usage{
mx.init.internal.default(name, shape, ctx, allow.unknown = FALSE)
}
\arguments{
\item{name}{the name of the variable.}

\item{shape}{the shape of the array to be generated.}
}
\description{
Internal default value initialization scheme.
}

