% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{mx.init.normal}
\alias{mx.init.normal}
\title{Create a initializer that initialize the weight with normal(0, sd)}
\usage{
mx.init.normal(sd)
}
\arguments{
\item{sd}{The standard deviation of normal distribution}
}
\description{
Create a initializer that initialize the weight with normal(0, sd)
}

