% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{mx.init.uniform}
\alias{mx.init.uniform}
\title{Create a initializer that initialize the weight with uniform [-scale, scale]}
\usage{
mx.init.uniform(scale)
}
\arguments{
\item{scale}{The scale of uniform distribution}
}
\description{
Create a initializer that initialize the weight with uniform [-scale, scale]
}

