% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{mx.init.Xavier}
\alias{mx.init.Xavier}
\title{Xavier initializer}
\usage{
mx.init.Xavier(rnd_type = "uniform", factor_type = "avg", magnitude = 3)
}
\arguments{
\item{rnd_type}{A string of \code{character} indicating the type of
distribution from which the weights are initialized.}

\item{factor_type}{A string of \code{character}.}

\item{magnitude}{A \code{numeric} number indicating the scale of random
number range.}
}
\description{
Create a initializer which initialize weight with Xavier or
similar initialization scheme.
}

