% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.io.CSVIter}
\alias{mx.io.CSVIter}
\title{Create iterator for dataset in csv.}
\usage{
mx.io.CSVIter(...)
}
\arguments{
\item{data.csv}{string, required
Dataset Param: Data csv path.}

\item{data.shape}{Shape(tuple), required
Dataset Param: Shape of the data.}

\item{label.csv}{string, optional, default='NULL'
Dataset Param: Label csv path. If is NULL, all labels will be returned as 0}

\item{label.shape}{Shape(tuple), optional, default=(1,)
Dataset Param: Shape of the label.}
}
\value{
iter The result mx.dataiter
}
\description{
Create iterator for dataset in csv.
}

