% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.io.ImageRecordUInt8Iter}
\alias{mx.io.ImageRecordUInt8Iter}
\title{Create iterator for dataset packed in recordio.}
\usage{
mx.io.ImageRecordUInt8Iter(...)
}
\arguments{
\item{path.imglist}{string, optional, default=''
Dataset Param: Path to image list.}

\item{path.imgrec}{string, optional, default='./data/imgrec.rec'
Dataset Param: Path to image record file.}

\item{aug.seq}{string, optional, default='aug_default'
Augmentation Param: the augmenter names to represent sequence of augmenters to be applied, seperated by comma. Additional keyword parameters will be seen by these augmenters.}

\item{label.width}{int, optional, default='1'
Dataset Param: How many labels for an image.}

\item{data.shape}{Shape(tuple), required
Dataset Param: Shape of each instance generated by the DataIter.}

\item{preprocess.threads}{int, optional, default='4'
Backend Param: Number of thread to do preprocessing.}

\item{verbose}{boolean, optional, default=True
Auxiliary Param: Whether to output parser information.}

\item{num.parts}{int, optional, default='1'
partition the data into multiple parts}

\item{part.index}{int, optional, default='0'
the index of the part will read}

\item{shuffle.chunk.size}{long (non-negative), optional, default=0
the size(MB) of the shuffle chunk, used with shuffle=True, it can enable global shuffling}

\item{shuffle.chunk.seed}{int, optional, default='0'
the seed for chunk shuffling}

\item{shuffle}{boolean, optional, default=False
Augmentation Param: Whether to shuffle data.}

\item{seed}{int, optional, default='0'
Augmentation Param: Random Seed.}

\item{batch.size}{int (non-negative), required
Batch Param: Batch size.}

\item{round.batch}{boolean, optional, default=True
Batch Param: Use round robin to handle overflow batch.}

\item{prefetch.buffer}{long (non-negative), optional, default=4
Backend Param: Number of prefetched parameters}

\item{dtype}{{'float16', 'float32', 'float64', 'invalid'},optional, default='invalid'
Data type.}

\item{resize}{int, optional, default='-1'
Augmentation Param: scale shorter edge to size before applying other augmentations.}

\item{rand.crop}{boolean, optional, default=False
Augmentation Param: Whether to random crop on the image}

\item{crop.y.start}{int, optional, default='-1'
Augmentation Param: Where to nonrandom crop on y.}

\item{crop.x.start}{int, optional, default='-1'
Augmentation Param: Where to nonrandom crop on x.}

\item{max.rotate.angle}{int, optional, default='0'
Augmentation Param: rotated randomly in [-max_rotate_angle, max_rotate_angle].}

\item{max.aspect.ratio}{float, optional, default=0
Augmentation Param: denotes the max ratio of random aspect ratio augmentation.}

\item{max.shear.ratio}{float, optional, default=0
Augmentation Param: denotes the max random shearing ratio.}

\item{max.crop.size}{int, optional, default='-1'
Augmentation Param: Maximum crop size.}

\item{min.crop.size}{int, optional, default='-1'
Augmentation Param: Minimum crop size.}

\item{max.random.scale}{float, optional, default=1
Augmentation Param: Maxmum scale ratio.}

\item{min.random.scale}{float, optional, default=1
Augmentation Param: Minimum scale ratio.}

\item{max.img.size}{float, optional, default=1e+10
Augmentation Param: Maxmum image size after resizing.}

\item{min.img.size}{float, optional, default=0
Augmentation Param: Minimum image size after resizing.}

\item{random.h}{int, optional, default='0'
Augmentation Param: Maximum value of H channel in HSL color space.}

\item{random.s}{int, optional, default='0'
Augmentation Param: Maximum value of S channel in HSL color space.}

\item{random.l}{int, optional, default='0'
Augmentation Param: Maximum value of L channel in HSL color space.}

\item{rotate}{int, optional, default='-1'
Augmentation Param: Rotate angle.}

\item{fill.value}{int, optional, default='255'
Augmentation Param: Maximum value of illumination variation.}

\item{data.shape}{Shape(tuple), required
Dataset Param: Shape of each instance generated by the DataIter.}

\item{inter.method}{int, optional, default='1'
Augmentation Param: 0-NN 1-bilinear 2-cubic 3-area 4-lanczos4 9-auto 10-rand.}

\item{pad}{int, optional, default='0'
Augmentation Param: Padding size.}
}
\value{
iter The result mx.dataiter
}
\description{
Create iterator for dataset packed in recordio.
}

