% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.io.MNISTIter}
\alias{mx.io.MNISTIter}
\title{Create iterator for MNIST hand-written digit number recognition dataset.}
\usage{
mx.io.MNISTIter(...)
}
\arguments{
\item{image}{string, optional, default='./train-images-idx3-ubyte'
Dataset Param: Mnist image path.}

\item{label}{string, optional, default='./train-labels-idx1-ubyte'
Dataset Param: Mnist label path.}

\item{batch.size}{int, optional, default='128'
Batch Param: Batch Size.}

\item{shuffle}{boolean, optional, default=True
Augmentation Param: Whether to shuffle data.}

\item{flat}{boolean, optional, default=False
Augmentation Param: Whether to flat the data into 1D.}

\item{seed}{int, optional, default='0'
Augmentation Param: Random Seed.}

\item{silent}{boolean, optional, default=False
Auxiliary Param: Whether to print out data info.}

\item{num.parts}{int, optional, default='1'
partition the data into multiple parts}

\item{part.index}{int, optional, default='0'
the index of the part will read}

\item{prefetch.buffer}{, optional, default=4
Backend Param: Number of prefetched parameters}
}
\value{
iter The result mx.dataiter
}
\description{
Create iterator for MNIST hand-written digit number recognition dataset.
}

