% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.accuracy}
\alias{mx.metric.accuracy}
\title{Accuracy metric}
\format{\preformatted{List of 3
 $ init  :function ()  
 $ update:function (label, pred, state)  
 $ get   :function (state)  
 - attr(*, "class")= chr "mx.metric"
}}
\usage{
mx.metric.accuracy
}
\description{
Accuracy metric
}
\keyword{datasets}

