% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.mse}
\alias{mx.metric.mse}
\title{MSE (Mean Squared Error) metric for regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.mse
}
\description{
MSE (Mean Squared Error) metric for regression
}
\keyword{datasets}

