% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.rmse}
\alias{mx.metric.rmse}
\title{RMSE (Root Mean Squared Error) metric for regression}
\format{\preformatted{List of 3
 $ init  :function ()  
 $ update:function (label, pred, state)  
 $ get   :function (state)  
 - attr(*, "class")= chr "mx.metric"
}}
\usage{
mx.metric.rmse
}
\description{
RMSE (Root Mean Squared Error) metric for regression
}
\keyword{datasets}

