% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{mx.model.FeedForward.create}
\alias{mx.model.FeedForward.create}
\title{Create a MXNet Feedforward neural net model with the specified training.}
\usage{
mx.model.FeedForward.create(symbol, X, y = NULL, ctx = NULL,
  num.round = 10, optimizer = "sgd", initializer = mx.init.uniform(0.01),
  eval.data = NULL, eval.metric = mx.metric.accuracy,
  iter.end.callback = NULL, epoch.end.callback = NULL,
  array.batch.size = 128, kvstore = "local", ...)
}
\arguments{
\item{symbol}{The symbolic configuration of the neural network.}

\item{X}{mx.io.DataIter or R array/matrix
The training data.}

\item{y}{R array, optional label of the data
This is only used when X is R array.}

\item{ctx}{mx.context or list of mx.context, optional
The devices used to perform training.}

\item{num.round}{integer (default=10)
The number of iterations over training data to train the model.}

\item{optimizer}{string, default="sgd"
The optimization method.}

\item{initializer,}{initializer object. default=mx.init.uniform(0.01)
The initialization scheme for parameters.}

\item{eval.data}{mx.io.DataIter or list(data=R.array, label=R.array), optional
The validation set used for validation evaluation during the progress}

\item{eval.metric}{function, optional
The evaluation function on the results.}

\item{iter.end.callback}{function, optional
The callback when iteration ends.}

\item{epoch.end.callback}{function, optional
The callback when one mini-batch iteration ends.}

\item{array.batch.size}{integer (default=128)
The batch size used for R array training.}

\item{kvstore}{string (default="local")
The parameter synchronization scheme in multiple devices.}
}
\value{
model A trained mxnet model.
}
\description{
Create a MXNet Feedforward neural net model with the specified training.
}

