% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mx.model.load}
\alias{mx.model.load}
\title{Load model checkpoint from file.}
\usage{
mx.model.load(prefix, iteration)
}
\arguments{
\item{prefix}{string prefix of the model name}

\item{iteration}{integer Iteration number of model we would like to load.}
}
\description{
Load model checkpoint from file.
}

