% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.argmax.channel}
\alias{mx.nd.argmax.channel}
\title{Take argmax indices of each channel of the src.The result will be ndarray of shape (num_channel,) on the same device.}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Take argmax indices of each channel of the src.The result will be ndarray of shape (num_channel,) on the same device.
}

