% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.argmin}
\alias{mx.nd.argmin}
\title{Compute argmin}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int, optional, default='-1'
Empty or unsigned. The axis to perform the reduction.If left empty, a global reduction will be performed.}

\item{keepdims}{boolean, optional, default=False
If true, the axis which is reduced is left in the result as dimension with size one.}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:74
}

