% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{mx.nd.array}
\alias{mx.nd.array}
\title{Create a new \code{mx.ndarray} that copies the content from src on ctx.}
\usage{
mx.nd.array(src.array, ctx = NULL)
}
\arguments{
\item{src.array}{Source array data of class \code{array}, \code{vector} or \code{matrix}.}

\item{ctx}{optional The context device of the array. mx.ctx.default() will be used in default.}
}
\value{
An \code{mx.ndarray}

An Rcpp_MXNDArray object
}
\description{
Create a new \code{mx.ndarray} that copies the content from src on ctx.
}
\examples{
mat = mx.nd.array(x)
mat = 1 - mat + (2 * mat)/(mat + 0.5)
as.array(mat)

}

