% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.batch.dot}
\alias{mx.nd.batch.dot}
\title{Calculate batched dot product of two matrices. (batch, M, K) batch_dot (batch, K, N) --> (batch, M, N)}
\arguments{
\item{lhs}{NDArray
Left operand  to the function}

\item{rhs}{NDArray
Right operand to the function}
}
\value{
out The result mx.ndarray
}
\description{
Calculate batched dot product of two matrices. (batch, M, K) batch_dot (batch, K, N) --> (batch, M, N)
}

