% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.BlockGrad}
\alias{mx.nd.BlockGrad}
\title{Get output from a symbol and pass 0 gradient back}
\arguments{
\item{data}{Symbol
Input data.}
}
\value{
out The result mx.ndarray
}
\description{
Get output from a symbol and pass 0 gradient back
}

