% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.axis}
\alias{mx.nd.broadcast.axis}
\title{Broadcast src along axis}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{Shape(tuple), optional, default=()
The axes to perform the broadcasting.}

\item{size}{Shape(tuple), optional, default=()
Target sizes of the broadcasting axes.}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:43
}

