% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.maximum}
\alias{mx.nd.broadcast.maximum}
\title{broadcast_maximum}
\arguments{
\item{lhs}{NDArray
first input}

\item{rhs}{NDArray
second input}
}
\value{
out The result mx.ndarray
}
\description{
broadcast_maximum
}

