% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.power}
\alias{mx.nd.broadcast.power}
\title{lhs power rhs with broadcast}
\arguments{
\item{lhs}{NDArray
Left operand  to the function}

\item{rhs}{NDArray
Right operand to the function}
}
\value{
out The result mx.ndarray
}
\description{
lhs power rhs with broadcast
}

