% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Cast}
\alias{mx.nd.Cast}
\title{Cast array to a different data type.}
\arguments{
\item{data}{Symbol
Input data to cast function.}

\item{dtype}{{'float16', 'float32', 'float64', 'int32', 'uint8'}, required
Target data type.}
}
\value{
out The result mx.ndarray
}
\description{
Cast array to a different data type.
}

