% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.choose.element.0index}
\alias{mx.nd.choose.element.0index}
\title{Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs. This function assume rhs uses 0-based index.}
\arguments{
\item{lhs}{NDArray
Left operand to the function.}

\item{rhs}{NDArray
Right operand to the function.}
}
\value{
out The result mx.ndarray
}
\description{
Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs. This function assume rhs uses 0-based index.
}

