% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.choose.element}
\alias{mx.nd.choose.element}
\title{Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs}
\arguments{
\item{lhs}{NDArray
Left operand to the function.}

\item{rhs}{NDArray
Right operand to the function.}
}
\value{
out The result mx.ndarray
}
\description{
Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs
}

