% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.clip}
\alias{mx.nd.clip}
\title{Clip ndarray elements to range (a_min, a_max)}
\arguments{
\item{src}{NDArray
Source input}

\item{a.min}{real_t
Minimum value}

\item{a.max}{real_t
Maximum value}
}
\value{
out The result mx.ndarray
}
\description{
Clip ndarray elements to range (a_min, a_max)
}

