% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Concat}
\alias{mx.nd.Concat}
\title{Perform an feature concat on channel dim (defaut is 1) over all}
\arguments{
\item{data}{Symbol[]
List of tensors to concatenate}

\item{num.args}{int, required
Number of inputs to be concated.}

\item{dim}{int, optional, default='1'
the dimension to be concated.}
}
\value{
out The result mx.ndarray
}
\description{
Perform an feature concat on channel dim (defaut is 1) over all
}

