% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Convolution}
\alias{mx.nd.Convolution}
\title{Apply convolution to input then add a bias.}
\arguments{
\item{data}{Symbol
Input data to the ConvolutionOp.}

\item{weight}{Symbol
Weight matrix.}

\item{bias}{Symbol
Bias parameter.}

\item{kernel}{Shape(tuple), required
convolution kernel size: (y, x) or (d, y, x)}

\item{stride}{Shape(tuple), optional, default=(1,1)
convolution stride: (y, x) or (d, y, x)}

\item{dilate}{Shape(tuple), optional, default=(1,1)
convolution dilate: (y, x)}

\item{pad}{Shape(tuple), optional, default=(0,0)
pad for convolution: (y, x) or (d, y, x)}

\item{num.filter}{int (non-negative), required
convolution filter(channel) number}

\item{num.group}{int (non-negative), optional, default=1
Number of groups partition. This option is not supported by CuDNN, you can use SliceChannel to num_group,apply convolution and concat instead to achieve the same need.}

\item{workspace}{long (non-negative), optional, default=1024
Tmp workspace for convolution (MB).}

\item{no.bias}{boolean, optional, default=False
Whether to disable bias parameter.}

\item{cudnn.tune}{{'fastest', 'limited_workspace', 'off'},optional, default='off'
Whether to find convolution algo by running performance test.Leads to higher startup time but may give better speed.auto tune is turned off by default.Set environment varialbe MXNET_CUDNN_AUTOTUNE_DEFAULT=1 to turn on by default.}

\item{cudnn.off}{boolean, optional, default=False
Turn off cudnn.}
}
\value{
out The result mx.ndarray
}
\description{
Apply convolution to input then add a bias.
}

