% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{mx.nd.copyto}
\alias{mx.nd.copyto}
\title{Generate an mx.ndarray object on ctx, with data copied from src}
\usage{
mx.nd.copyto(src, ctx)
}
\arguments{
\item{src}{The source mx.ndarray object.}

\item{ctx}{The target context.}
}
\description{
Generate an mx.ndarray object on ctx, with data copied from src
}

