% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Custom}
\alias{mx.nd.Custom}
\title{Custom operator implemented in frontend.}
\arguments{
\item{op.type}{string
Type of custom operator. Must be registered first.}
}
\value{
out The result mx.ndarray
}
\description{
Custom operator implemented in frontend.
}

