% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.degrees}
\alias{mx.nd.degrees}
\title{Take degrees of the src}
\arguments{
\item{src}{NDArray
Source input}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:230
}

