% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.dot}
\alias{mx.nd.dot}
\title{Calculate dot product of two matrices or two vectors.}
\arguments{
\item{lhs}{NDArray
Left input}

\item{rhs}{NDArray
Right input}

\item{axis}{int, required
The dimension to flip}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/matrix_op.cc:154
}

