% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Dropout}
\alias{mx.nd.Dropout}
\title{Apply dropout to input.
During training, each element of the input is randomly set to zero with probability p.
And then the whole tensor is rescaled by 1/(1-p) to keep the expectation the same as
before applying dropout. During the test time, this behaves as an identity map.}
\arguments{
\item{data}{Symbol
Input data to dropout.}

\item{p}{float, optional, default=0.5
Fraction of the input that gets dropped out at training time}
}
\value{
out The result mx.ndarray
}
\description{
Apply dropout to input.
During training, each element of the input is randomly set to zero with probability p.
And then the whole tensor is rescaled by 1/(1-p) to keep the expectation the same as
before applying dropout. During the test time, this behaves as an identity map.
}

