% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Embedding}
\alias{mx.nd.Embedding}
\title{Map integer index to vector representations (embeddings). Those
embeddings are learnable parameters. For a input of shape `(d1, ..., dK)`, the
output shape is `(d1, ..., dK, output_dim)`. All the input values should be
integers in the range `[0, input_dim)`.}
\arguments{
\item{data}{Symbol
Input data to the EmbeddingOp.}

\item{weight}{Symbol
Enbedding weight matrix.}

\item{input.dim}{int, required
vocabulary size of the input indices.}

\item{output.dim}{int, required
dimension of the embedding vectors.}
}
\value{
out The result mx.ndarray
}
\description{
Map integer index to vector representations (embeddings). Those
embeddings are learnable parameters. For a input of shape `(d1, ..., dK)`, the
output shape is `(d1, ..., dK, output_dim)`. All the input values should be
integers in the range `[0, input_dim)`.
}

