% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.fill.element.0index}
\alias{mx.nd.fill.element.0index}
\title{Fill one element of each line(row for python, column for R/Julia) in lhs according to index indicated by rhs and values indicated by mhs. This function assume rhs uses 0-based index.}
\arguments{
\item{lhs}{NDArray
Left operand to the function.}

\item{mhs}{NDArray
Middle operand to the function.}

\item{rhs}{NDArray
Right operand to the function.}
}
\value{
out The result mx.ndarray
}
\description{
Fill one element of each line(row for python, column for R/Julia) in lhs according to index indicated by rhs and values indicated by mhs. This function assume rhs uses 0-based index.
}

