% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Flatten}
\alias{mx.nd.Flatten}
\title{Flatten input into 2D by collapsing all the higher dimensions.
A (d1, d2, ..., dK) tensor is flatten to (d1, d2* ... *dK) matrix.}
\arguments{
\item{data}{Symbol
Input data to flatten.}
}
\value{
out The result mx.ndarray
}
\description{
Flatten input into 2D by collapsing all the higher dimensions.
A (d1, d2, ..., dK) tensor is flatten to (d1, d2* ... *dK) matrix.
}

