% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.FullyConnected}
\alias{mx.nd.FullyConnected}
\title{Apply matrix multiplication to input then add a bias.
It maps the input of shape `(batch_size, input_dim)` to the shape of
`(batch_size, num_hidden)`. Learnable parameters include the weights
of the linear transform and an optional bias vector.}
\arguments{
\item{data}{Symbol
Input data to the FullyConnectedOp.}

\item{weight}{Symbol
Weight matrix.}

\item{bias}{Symbol
Bias parameter.}

\item{num.hidden}{int, required
Number of hidden nodes of the output.}

\item{no.bias}{boolean, optional, default=False
Whether to disable bias parameter.}
}
\value{
out The result mx.ndarray
}
\description{
Apply matrix multiplication to input then add a bias.
It maps the input of shape `(batch_size, input_dim)` to the shape of
`(batch_size, num_hidden)`. Learnable parameters include the weights
of the linear transform and an optional bias vector.
}

