% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.identity}
\alias{mx.nd.identity}
\title{Identity mapping, copy src to output}
\arguments{
\item{src}{NDArray
Source input}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:14
}

