% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.LeakyReLU}
\alias{mx.nd.LeakyReLU}
\title{Apply activation function to input.}
\arguments{
\item{data}{Symbol
Input data to activation function.}

\item{act.type}{{'elu', 'leaky', 'prelu', 'rrelu'},optional, default='leaky'
Activation function to be applied.}

\item{slope}{float, optional, default=0.25
Init slope for the activation. (For leaky and elu only)}

\item{lower.bound}{float, optional, default=0.125
Lower bound of random slope. (For rrelu only)}

\item{upper.bound}{float, optional, default=0.334
Upper bound of random slope. (For rrelu only)}
}
\value{
out The result mx.ndarray
}
\description{
Apply activation function to input.
}

