% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.LinearRegressionOutput}
\alias{mx.nd.LinearRegressionOutput}
\title{Use linear regression for final output, this is used on final output of a net.}
\arguments{
\item{data}{Symbol
Input data to function.}

\item{label}{Symbol
Input label to function.}

\item{grad.scale}{float, optional, default=1
Scale the gradient by a float factor}
}
\value{
out The result mx.ndarray
}
\description{
Use linear regression for final output, this is used on final output of a net.
}

