% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{mx.nd.load}
\alias{mx.nd.load}
\title{Load an mx.nd.array object on disk}
\usage{
mx.nd.load(filename)
}
\arguments{
\item{filename}{the filename (including the path)}
}
\description{
Load an mx.nd.array object on disk
}
\examples{
mat = mx.nd.array(1:3)
mx.nd.save(mat, 'temp.mat')
mat2 = mx.nd.load('temp.mat')
as.array(mat)
as.array(mat2)

}

