% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.log1p}
\alias{mx.nd.log1p}
\title{Take `log(1 + x)` in a numerically stable way}
\arguments{
\item{src}{NDArray
Source input}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:167
}

