% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.log2}
\alias{mx.nd.log2}
\title{Take base-2 log of the src}
\arguments{
\item{src}{NDArray
Source input}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:149
}

