% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.LRN}
\alias{mx.nd.LRN}
\title{Apply convolution to input then add a bias.}
\arguments{
\item{data}{Symbol
Input data to the ConvolutionOp.}

\item{alpha}{float, optional, default=0.0001
value of the alpha variance scaling parameter in the normalization formula}

\item{beta}{float, optional, default=0.75
value of the beta power parameter in the normalization formula}

\item{knorm}{float, optional, default=2
value of the k parameter in normalization formula}

\item{nsize}{int (non-negative), required
normalization window width in elements.}
}
\value{
out The result mx.ndarray
}
\description{
Apply convolution to input then add a bias.
}

