% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.MAERegressionOutput}
\alias{mx.nd.MAERegressionOutput}
\title{Use mean absolute error regression for final output, this is used on final output of a net.}
\arguments{
\item{data}{Symbol
Input data to function.}

\item{label}{Symbol
Input label to function.}

\item{grad.scale}{float, optional, default=1
Scale the gradient by a float factor}
}
\value{
out The result mx.ndarray
}
\description{
Use mean absolute error regression for final output, this is used on final output of a net.
}

