% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.norm}
\alias{mx.nd.norm}
\title{Take L2 norm of the src.The result will be ndarray of shape (1,) on the same device.}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Take L2 norm of the src.The result will be ndarray of shape (1,) on the same device.
}

