% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.normal}
\alias{mx.nd.normal}
\title{Sample a normal distribution}
\arguments{
\item{loc}{float, optional, default=0
Mean of the distribution.}

\item{scale}{float, optional, default=1
Standard deviation of the distribution.}

\item{shape}{Shape(tuple), optional, default=()
The shape of the output}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n).Only used for imperative calls.}
}
\value{
out The result mx.ndarray
}
\description{
Sample a normal distribution
}

