% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Pad}
\alias{mx.nd.Pad}
\title{Pads an n-dimensional input tensor. Allows for precise control of the padding type and how much padding to apply on both sides of a given dimension.}
\arguments{
\item{data}{Symbol
An n-dimensional input tensor.}

\item{mode}{{'constant', 'edge'}, required
Padding type to use. "constant" pads all values with a constant value, the value of which can be specified with the constant_value option. "edge" uses the boundary values of the array as padding.}

\item{pad.width}{Shape(tuple), required
A tuple of padding widths of length 2*r, where r is the rank of the input tensor, specifying number of values padded to the edges of each axis. (before_1, after_1, ... , before_N, after_N) unique pad widths for each axis. Equivalent to pad_width in numpy.pad, but flattened.}

\item{constant.value}{double, optional, default=0
This option is only used when mode is "constant". This value will be used as the padding value. Defaults to 0 if not specified.}
}
\value{
out The result mx.ndarray
}
\description{
Pads an n-dimensional input tensor. Allows for precise control of the padding type and how much padding to apply on both sides of a given dimension.
}

