% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.ROIPooling}
\alias{mx.nd.ROIPooling}
\title{Performs region-of-interest pooling on inputs. Resize bounding box coordinates by spatial_scale and crop input feature maps accordingly. The cropped feature maps are pooled by max pooling to a fixed size output indicated by pooled_size. batch_size will change to the number of region bounding boxes after ROIPooling}
\arguments{
\item{data}{Symbol
Input data to the pooling operator, a 4D Feature maps}

\item{rois}{Symbol
Bounding box coordinates, a 2D array of [[batch_index, x1, y1, x2, y2]]. (x1, y1) and (x2, y2) are top left and down right corners of designated region of interest. batch_index indicates the index of corresponding image in the input data}

\item{pooled.size}{Shape(tuple), required
fix pooled size: (h, w)}

\item{spatial.scale}{float, required
Ratio of input feature map height (or w) to raw image height (or w). Equals the reciprocal of total stride in convolutional layers}
}
\value{
out The result mx.ndarray
}
\description{
Performs region-of-interest pooling on inputs. Resize bounding box coordinates by spatial_scale and crop input feature maps accordingly. The cropped feature maps are pooled by max pooling to a fixed size output indicated by pooled_size. batch_size will change to the number of region bounding boxes after ROIPooling
}

