% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.rsqrt}
\alias{mx.nd.rsqrt}
\title{Take rsqrt of the src}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Take rsqrt of the src
}

