% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.slice.axis}
\alias{mx.nd.slice.axis}
\title{Slice the input along certain axis and return a sliced array.}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int, required
The axis to be sliced}

\item{begin}{int, required
The beginning index to be sliced}

\item{end}{int, required
The end index to be sliced}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/matrix_op.cc:120
}

