% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.SliceChannel}
\alias{mx.nd.SliceChannel}
\title{Slice input equally along specified axis}
\arguments{
\item{num.outputs}{int, required
Number of outputs to be sliced.}

\item{axis}{int, optional, default='1'
Dimension along which to slice.}

\item{squeeze.axis}{boolean, optional, default=False
If true AND the sliced dimension becomes 1, squeeze that dimension.}
}
\value{
out The result mx.ndarray
}
\description{
Slice input equally along specified axis
}

