% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.smooth.l1}
\alias{mx.nd.smooth.l1}
\title{Calculate Smooth L1 Loss(lhs, scalar)}
\arguments{
\item{lhs}{NDArray
source input}

\item{scalar}{float
scalar input}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/elemwise_binary_scalar_op.cc:98
}

