% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.softmax.cross.entropy}
\alias{mx.nd.softmax.cross.entropy}
\title{Calculate cross_entropy(lhs, one_hot(rhs))}
\arguments{
\item{data}{NDArray
Input data}

\item{label}{NDArray
Input label}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/loss_binary_op.cc:12
}

