% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sqrt}
\alias{mx.nd.sqrt}
\title{Take square root of the src}
\arguments{
\item{src}{NDArray
Source input to the function.}
}
\value{
out The result mx.ndarray
}
\description{
Take square root of the src
}

