% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sum.mid.internal}
\alias{mx.nd.sum.mid.internal}
\title{Take sum on medium dimension of the 3D src.}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Take sum on medium dimension of the 3D src.
}

