% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sum}
\alias{mx.nd.sum}
\title{Sum src along axis. If axis is empty, global reduction is performed}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{Shape(tuple), optional, default=()
Empty or unsigned or tuple. The axes to perform the reduction.If left empty, a global reduction will be performed.}

\item{keepdims}{boolean, optional, default=False
If true, the axis which is reduced is left in the result as dimension with size one.}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:16
}

