% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.SVMOutput}
\alias{mx.nd.SVMOutput}
\title{Support Vector Machine based transformation on input, backprop L2-SVM}
\arguments{
\item{data}{Symbol
Input data to svm.}

\item{label}{Symbol
Label data.}

\item{margin}{float, optional, default=1
Scale the DType(param_.margin) for activation size}

\item{regularization.coefficient}{float, optional, default=1
Scale the coefficient responsible for balacing coefficient size and error tradeoff}

\item{use.linear}{boolean, optional, default=False
If set true, uses L1-SVM objective function. Default uses L2-SVM objective}
}
\value{
out The result mx.ndarray
}
\description{
Support Vector Machine based transformation on input, backprop L2-SVM
}

