% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.SwapAxis}
\alias{mx.nd.SwapAxis}
\title{Apply swapaxis to input.}
\arguments{
\item{data}{Symbol
Input data to the SwapAxisOp.}

\item{dim1}{int (non-negative), optional, default=0
the first axis to be swapped.}

\item{dim2}{int (non-negative), optional, default=0
the second axis to be swapped.}
}
\value{
out The result mx.ndarray
}
\description{
Apply swapaxis to input.
}

