% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.transpose}
\alias{mx.nd.transpose}
\title{Transpose the input matrix and return a new one}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Transpose the input matrix and return a new one
}

