% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.uniform}
\alias{mx.nd.uniform}
\title{Sample a uniform distribution}
\arguments{
\item{low}{float, optional, default=0
The lower bound of distribution}

\item{high}{float, optional, default=1
The upper bound of distribution}

\item{shape}{Shape(tuple), optional, default=()
The shape of the output}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n).Only used for imperative calls.}
}
\value{
out The result mx.ndarray
}
\description{
Sample a uniform distribution
}

