% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mx.opt.adadelta}
\alias{mx.opt.adadelta}
\title{Create an AdaDelta optimizer with respective parameters.}
\usage{
mx.opt.adadelta(rho = 0.9, epsilon = 1e-05, wd = 0, rescale.grad = 1,
  clip_gradient = NULL)
}
\arguments{
\item{rho}{float, default=0.90
Decay rate for both squared gradients and delta x.}

\item{epsilon}{float, default=1e-5
The constant as described in the thesis.}

\item{wd}{float, default=0.0
L2 regularization coefficient add to all the weights.}

\item{rescale.grad}{float, default=1.0
rescaling factor of gradient.}

\item{clip_gradient}{float, optional
clip gradient in range [-clip_gradient, clip_gradient].}
}
\description{
AdaDelta optimizer as described in Zeiler, M. D. (2012).
*ADADELTA: An adaptive learning rate method.*
http://arxiv.org/abs/1212.5701
}

