% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mx.opt.create}
\alias{mx.opt.create}
\title{Create an optimizer by name and parameters}
\usage{
mx.opt.create(name, ...)
}
\arguments{
\item{name}{The name of the optimizer}

\item{...}{Additional arguments}
}
\description{
Create an optimizer by name and parameters
}

