% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mx.opt.get.updater}
\alias{mx.opt.get.updater}
\title{Get an updater closure that can take list of weight and gradient
and return updated list of weight.}
\usage{
mx.opt.get.updater(optimizer, weights)
}
\arguments{
\item{optimizer}{The optimizer}

\item{weights}{The weights to be optimized}
}
\description{
Get an updater closure that can take list of weight and gradient
and return updated list of weight.
}

