% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mx.opt.sgd}
\alias{mx.opt.sgd}
\title{Create an SGD optimizer with respective parameters.
Perform SGD with momentum update}
\usage{
mx.opt.sgd(learning.rate, momentum = 0, wd = 0, rescale.grad = 1,
  clip_gradient = NULL, lr_scheduler = NULL)
}
\description{
Create an SGD optimizer with respective parameters.
Perform SGD with momentum update
}

