% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.R
\name{mx.rnn.forward}
\alias{mx.rnn.forward}
\title{Using forward function to predict in rnn inference model}
\usage{
mx.rnn.forward(model, input.data, new.seq = FALSE)
}
\arguments{
\item{model}{rnn model
A rnn inference model}

\item{input.data, }{array.matrix
The input data for forward function}

\item{new.seq}{boolean, default=FALSE
Whether the input is the start of a new sequence}
}
\value{
result A list(prob=prob, model=model) containing the result probability of each label and the model.
}
\description{
Using forward function to predict in rnn inference model
}

