% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.R
\name{mx.rnn.inference}
\alias{mx.rnn.inference}
\title{Create a RNN Inference Model}
\usage{
mx.rnn.inference(num.rnn.layer, input.size, num.hidden, num.embed, num.label,
  batch.size = 1, arg.params, ctx = mx.cpu(), dropout = 0,
  batch.norm = FALSE)
}
\arguments{
\item{num.rnn.layer}{integer
The number of the layer of rnn.}

\item{input.size}{integer
The input dim of one-hot encoding of embedding}

\item{num.hidden}{integer
The number of hidden nodes.}

\item{num.embed}{integer
The output dim of embedding.}

\item{num.label}{integer
The number of labels.}

\item{batch.size}{integer, default=1
The batch size used for R array training.}

\item{arg.params}{list
The batch size used for R array training.}

\item{ctx}{mx.context, optional
Model parameter, list of name to NDArray of net's weights.}

\item{dropout}{float, default=0
A number in [0,1) containing the dropout ratio from the last hidden layer to the output layer.}

\item{batch.norm}{boolean, default=FALSE
Whether to use batch normalization.}
}
\value{
model list(rnn.exec=integer, symbol=mxnet symbol, num.rnn.layer=integer, num.hidden=integer, seq.len=integer, batch.size=integer, num.embed=integer) 
     A rnn inference model.
}
\description{
Create a RNN Inference Model
}

