% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mx.runif}
\alias{mx.runif}
\title{Generate uniform distribution in [low, high) with specified shape.}
\usage{
mx.runif(shape, min = 0, max = 1, ctx = NULL)
}
\arguments{
\item{shape}{Dimension, The shape(dimension) of the result.}

\item{min}{numeric, The lower bound of distribution.}

\item{max}{numeric, The upper bound of distribution.}

\item{ctx, }{optional The context device of the array. mx.ctx.default() will be used in default.}
}
\description{
Generate uniform distribution in [low, high) with specified shape.
}
\examples{

mx.set.seed(0)
as.array(mx.runif(2))
# 0.5488135 0.5928446
mx.set.seed(0)
as.array(mx.rnorm(2))
# 2.212206 1.163079

}

