% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executor.R
\name{mx.simple.bind}
\alias{mx.simple.bind}
\title{Simple bind the symbol to executor,
with information from input shapes.}
\usage{
mx.simple.bind(symbol, ctx, grad.req = "null", ...)
}
\description{
Simple bind the symbol to executor,
with information from input shapes.
}

