% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.adam_update}
\alias{mx.symbol.adam_update}
\title{adam_update:Updater function for adam optimizer}
\usage{
mx.symbol.adam_update(...)
}
\arguments{
\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
adam_update:Updater function for adam optimizer
}

