% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.argmax_channel}
\alias{mx.symbol.argmax_channel}
\title{argmax_channel:}
\usage{
mx.symbol.argmax_channel(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
argmax_channel:
}

