% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.argmin}
\alias{mx.symbol.argmin}
\title{argmin:Compute argmin}
\usage{
mx.symbol.argmin(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int, optional, default='-1'
Empty or unsigned. The axis to perform the reduction.If left empty, a global reduction will be performed.}

\item{keepdims}{boolean, optional, default=False
If true, the axis which is reduced is left in the result as dimension with size one.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:74
}

