% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.BlockGrad}
\alias{mx.symbol.BlockGrad}
\title{Get output from a symbol and pass 0 gradient back}
\usage{
mx.symbol.BlockGrad(...)
}
\arguments{
\item{data}{Symbol
Input data.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Get output from a symbol and pass 0 gradient back
}

