% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_axis}
\alias{mx.symbol.broadcast_axis}
\title{Broadcast data in the given axis to the given size. The original size of the broadcasting axis must be 1.}
\usage{
mx.symbol.broadcast_axis(...)
}
\arguments{
\item{src}{Symbol
Left symbolic input to the function}

\item{axis}{Shape(tuple), optional, default=()
The axes to perform the broadcasting.}

\item{size}{Shape(tuple), optional, default=()
Target sizes of the broadcasting axes.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Broadcast data in the given axis to the given size. The original size of the broadcasting axis must be 1.
}

