% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_minus}
\alias{mx.symbol.broadcast_minus}
\title{lhs minus rhs with broadcast}
\usage{
mx.symbol.broadcast_minus(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Right symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
lhs minus rhs with broadcast
}

