% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_power}
\alias{mx.symbol.broadcast_power}
\title{lhs power rhs with broadcast}
\usage{
mx.symbol.broadcast_power(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Right symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
lhs power rhs with broadcast
}

