% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_to}
\alias{mx.symbol.broadcast_to}
\title{broadcast_to:Broadcast src to shape}
\usage{
mx.symbol.broadcast_to(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{shape}{Shape(tuple), optional, default=()
The shape of the desired array. We can set the dim to zero if it's same as the original. E.g `A = broadcast_to(B, shape=(10, 0, 0))` has the same meaning as `A = broadcast_axis(B, axis=0, size=10)`.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:50
}

