% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Cast}
\alias{mx.symbol.Cast}
\title{Cast array to a different data type.}
\usage{
mx.symbol.Cast(...)
}
\arguments{
\item{data}{Symbol
Input data to cast function.}

\item{dtype}{{'float16', 'float32', 'float64', 'int32', 'uint8'}, required
Target data type.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Cast array to a different data type.
}

