% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.choose_element_0index}
\alias{mx.symbol.choose_element_0index}
\title{choose_element_0index:Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs. This function assume rhs uses 0-based index.}
\usage{
mx.symbol.choose_element_0index(...)
}
\arguments{
\item{lhs}{NDArray
Left operand to the function.}

\item{rhs}{NDArray
Right operand to the function.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
choose_element_0index:Choose one element from each line(row for python, column for R/Julia) in lhs according to index indicated by rhs. This function assume rhs uses 0-based index.
}

