% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.clip}
\alias{mx.symbol.clip}
\title{clip:Clip ndarray elements to range (a_min, a_max)}
\usage{
mx.symbol.clip(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{a.min}{real_t
Minimum value}

\item{a.max}{real_t
Maximum value}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
clip:Clip ndarray elements to range (a_min, a_max)
}

