% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Correlation}
\alias{mx.symbol.Correlation}
\title{Correlation:Apply correlation to inputs}
\usage{
mx.symbol.Correlation(...)
}
\arguments{
\item{data1}{Symbol
Input data1 to the correlation.}

\item{data2}{Symbol
Input data2 to the correlation.}

\item{kernel.size}{int (non-negative), optional, default=1
kernel size for Correlation must be an odd number}

\item{max.displacement}{int (non-negative), optional, default=1
Max displacement of Correlation}

\item{stride1}{int (non-negative), optional, default=1
stride1 quantize data1 globally}

\item{stride2}{int (non-negative), optional, default=1
stride2 quantize data2 within the neighborhood centered around data1}

\item{pad.size}{int (non-negative), optional, default=0
pad for Correlation}

\item{is.multiply}{boolean, optional, default=True
operation type is either multiplication or subduction}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Correlation:Apply correlation to inputs
}

