% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.crop}
\alias{mx.symbol.crop}
\title{crop:(Crop the input tensor and return a new one.}
\usage{
mx.symbol.crop(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{begin}{Shape(tuple), required
starting coordinates}

\item{end}{Shape(tuple), required
ending coordinates}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Requirements
------------
- the input and output (if explicitly given) are of the same data type,
  and on the same device.
)
}
\details{
From:src/operator/tensor/matrix_op.cc:69
}

