% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Custom}
\alias{mx.symbol.Custom}
\title{Custom:Custom operator implemented in frontend.}
\usage{
mx.symbol.Custom(...)
}
\arguments{
\item{op.type}{string
Type of custom operator. Must be registered first.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Custom:Custom operator implemented in frontend.
}

