% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Deconvolution}
\alias{mx.symbol.Deconvolution}
\title{Apply deconvolution to input then add a bias.}
\usage{
mx.symbol.Deconvolution(...)
}
\arguments{
\item{data}{Symbol
Input data to the DeconvolutionOp.}

\item{weight}{Symbol
Weight matrix.}

\item{bias}{Symbol
Bias parameter.}

\item{kernel}{Shape(tuple), required
deconvolution kernel size: (y, x)}

\item{stride}{Shape(tuple), optional, default=(1,1)
deconvolution stride: (y, x)}

\item{pad}{Shape(tuple), optional, default=(0,0)
pad for deconvolution: (y, x)}

\item{num.filter}{int (non-negative), required
deconvolution filter(channel) number}

\item{num.group}{int (non-negative), optional, default=1
number of groups partition}

\item{workspace}{long (non-negative), optional, default=512
Tmp workspace for deconvolution (MB)}

\item{no.bias}{boolean, optional, default=False
Whether to disable bias parameter.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply deconvolution to input then add a bias.
}

