% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.dot}
\alias{mx.symbol.dot}
\title{Calculate dot product of two matrices or two vectors}
\usage{
mx.symbol.dot(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Right symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Calculate dot product of two matrices or two vectors
}

