% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Dropout}
\alias{mx.symbol.Dropout}
\title{Dropout:Apply dropout to input.
During training, each element of the input is randomly set to zero with probability p.
And then the whole tensor is rescaled by 1/(1-p) to keep the expectation the same as
before applying dropout. During the test time, this behaves as an identity map.}
\usage{
mx.symbol.Dropout(...)
}
\arguments{
\item{data}{Symbol
Input data to dropout.}

\item{p}{float, optional, default=0.5
Fraction of the input that gets dropped out at training time}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Dropout:Apply dropout to input.
During training, each element of the input is randomly set to zero with probability p.
And then the whole tensor is rescaled by 1/(1-p) to keep the expectation the same as
before applying dropout. During the test time, this behaves as an identity map.
}

