% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.ElementWiseSum}
\alias{mx.symbol.ElementWiseSum}
\title{ElementWiseSum:Perform element sum of inputs}
\usage{
mx.symbol.ElementWiseSum(...)
}
\arguments{
\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_sum.cc:49
}

