% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.expm1}
\alias{mx.symbol.expm1}
\title{expm1:Take `exp(x) - 1` in a numerically stable way}
\usage{
mx.symbol.expm1(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:176
}

