% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.fix}
\alias{mx.symbol.fix}
\title{fix:Take round of the src to integer nearest 0}
\usage{
mx.symbol.fix(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:98
}

