% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.flip}
\alias{mx.symbol.flip}
\title{flip:Flip the input tensor along axis and return a new one.}
\usage{
mx.symbol.flip(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int, required
The dimension to flip}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/matrix_op.cc:142
}

