% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.gammaln}
\alias{mx.symbol.gammaln}
\title{gammaln:Take gammaln (log of the absolute value of gamma(x)) of the src}
\usage{
mx.symbol.gammaln(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:311
}

