% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mx.symbol.Group}
\alias{mx.symbol.Group}
\title{Create a symbol that groups symbols together.}
\usage{
mx.symbol.Group(...)
}
\arguments{
\item{kwarg}{Variable length of symbols or list of symbol.}
}
\value{
The result symbol
}
\description{
Create a symbol that groups symbols together.
}

