% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.IdentityAttachKLSparseReg}
\alias{mx.symbol.IdentityAttachKLSparseReg}
\title{Apply a sparse regularization to the output a sigmoid activation function.}
\usage{
mx.symbol.IdentityAttachKLSparseReg(...)
}
\arguments{
\item{data}{Symbol
Input data.}

\item{sparseness.target}{float, optional, default=0.1
The sparseness target}

\item{penalty}{float, optional, default=0.001
The tradeoff parameter for the sparseness penalty}

\item{momentum}{float, optional, default=0.9
The momentum for running average}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply a sparse regularization to the output a sigmoid activation function.
}

