% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.InstanceNorm}
\alias{mx.symbol.InstanceNorm}
\title{InstanceNorm:An operator taking in a n-dimensional input tensor (n > 2), and normalizing the input by subtracting the mean and variance calculated over the spatial dimensions. This is an implemention of the operator described in "Instance Normalization: The Missing Ingredient for Fast Stylization", D. Ulyanov, A. Vedaldi, V. Lempitsky, 2016 (arXiv:1607.08022v2). This layer is similar to batch normalization, with two differences: first, the normalization is carried out per example ('instance'), not over a batch. Second, the same normalization is applied both at test and train time. This operation is also known as 'contrast normalization'.}
\usage{
mx.symbol.InstanceNorm(...)
}
\arguments{
\item{data}{Symbol
A n-dimensional tensor (n > 2) of the form [batch, channel, spatial_dim1, spatial_dim2, ...].}

\item{gamma}{Symbol
A vector of length 'channel', which multiplies the normalized input.}

\item{beta}{Symbol
A vector of length 'channel', which is added to the product of the normalized input and the weight.}

\item{eps}{float, optional, default=0.001
Epsilon to prevent division by 0.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
InstanceNorm:An operator taking in a n-dimensional input tensor (n > 2), and normalizing the input by subtracting the mean and variance calculated over the spatial dimensions. This is an implemention of the operator described in "Instance Normalization: The Missing Ingredient for Fast Stylization", D. Ulyanov, A. Vedaldi, V. Lempitsky, 2016 (arXiv:1607.08022v2). This layer is similar to batch normalization, with two differences: first, the normalization is carried out per example ('instance'), not over a batch. Second, the same normalization is applied both at test and train time. This operation is also known as 'contrast normalization'.
}

