% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.L2Normalization}
\alias{mx.symbol.L2Normalization}
\title{Set the l2 norm of each instance to a constant.}
\usage{
mx.symbol.L2Normalization(...)
}
\arguments{
\item{data}{Symbol
Input data to the L2NormalizationOp.}

\item{eps}{float, optional, default=1e-10
Epsilon to prevent div 0}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Set the l2 norm of each instance to a constant.
}

