% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.LinearRegressionOutput}
\alias{mx.symbol.LinearRegressionOutput}
\title{Use linear regression for final output, this is used on final output of a net.}
\usage{
mx.symbol.LinearRegressionOutput(...)
}
\arguments{
\item{data}{Symbol
Input data to function.}

\item{label}{Symbol
Input label to function.}

\item{grad.scale}{float, optional, default=1
Scale the gradient by a float factor}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Use linear regression for final output, this is used on final output of a net.
}

