% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mx.symbol.load.json}
\alias{mx.symbol.load.json}
\title{Load an mx.symbol object from a json string}
\arguments{
\item{str}{the json str represent a mx.symbol}
}
\description{
Load an mx.symbol object from a json string
}

