% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mx.symbol.load}
\alias{mx.symbol.load}
\title{Load an mx.symbol object}
\usage{
mx.symbol.load(filename)
}
\arguments{
\item{filename}{the filename (including the path)}
}
\description{
Load an mx.symbol object
}
\examples{
data = mx.symbol.Variable('data')
mx.symbol.save(data, 'temp.symbol')
data2 = mx.symbol.load('temp.symbol')

}

