% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.log1p}
\alias{mx.symbol.log1p}
\title{log1p:Take `log(1 + x)` in a numerically stable way}
\usage{
mx.symbol.log1p(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:167
}

