% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.log2}
\alias{mx.symbol.log2}
\title{log2:Take base-2 log of the src}
\usage{
mx.symbol.log2(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:149
}

