% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.negative}
\alias{mx.symbol.negative}
\title{negative:Negate src}
\usage{
mx.symbol.negative(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:54
}

