% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.norm}
\alias{mx.symbol.norm}
\title{norm:}
\usage{
mx.symbol.norm(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
norm:
}

