% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.normal}
\alias{mx.symbol.normal}
\title{normal:Sample a normal distribution}
\usage{
mx.symbol.normal(...)
}
\arguments{
\item{loc}{float, optional, default=0
Mean of the distribution.}

\item{scale}{float, optional, default=1
Standard deviation of the distribution.}

\item{shape}{Shape(tuple), optional, default=()
The shape of the output}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n).Only used for imperative calls.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
normal:Sample a normal distribution
}

