% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Pooling}
\alias{mx.symbol.Pooling}
\title{Pooling:Perform spatial pooling on inputs.}
\usage{
mx.symbol.Pooling(...)
}
\arguments{
\item{data}{Symbol
Input data to the pooling operator.}

\item{global.pool}{boolean, optional, default=False
Ignore kernel size, do global pooling based on current input feature map. This is useful for input with different shape}

\item{kernel}{Shape(tuple), required
pooling kernel size: (y, x) or (d, y, x)}

\item{pool.type}{{'avg', 'max', 'sum'}, required
Pooling type to be applied.}

\item{pooling.convention}{{'full', 'valid'},optional, default='valid'
Pooling convention to be applied.kValid is default setting of Mxnet and rounds down the output pooling size.kFull is compatible with Caffe and rounds up the output pooling size.}

\item{stride}{Shape(tuple), optional, default=(1,1)
stride: for pooling (y, x) or (d, y, x)}

\item{pad}{Shape(tuple), optional, default=(0,0)
pad for pooling: (y, x) or (d, y, x)}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Pooling:Perform spatial pooling on inputs.
}

