% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.ROIPooling}
\alias{mx.symbol.ROIPooling}
\title{Resize regions of interest in an input plane to a fixed size by MAX pooling.}
\usage{
mx.symbol.ROIPooling(...)
}
\arguments{
\item{data}{Symbol[]
[input tensor, regions of interest]}

\item{pooled.size}{Shape(tuple), required
target size: (h, w)}

\item{spatial.scale}{float, required
Ratio of input plane height (or w) to raw image height (or w).}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Resize regions of interest in an input plane to a fixed size by MAX pooling.
}

