% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.rsqrt}
\alias{mx.symbol.rsqrt}
\title{rsqrt:Take reciprocal square root of the src}
\usage{
mx.symbol.rsqrt(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_unary_op.cc:121
}

