% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mx.symbol.save}
\alias{mx.symbol.save}
\title{Save an mx.symbol object}
\usage{
mx.symbol.save(symbol, filename)
}
\arguments{
\item{symbol}{the \code{mx.symbol} object}

\item{filename}{the filename (including the path)}
}
\description{
Save an mx.symbol object
}
\examples{
data = mx.symbol.Variable('data')
mx.symbol.save(data, 'temp.symbol')
data2 = mx.symbol.load('temp.symbol')

}

