% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SequenceReverse}
\alias{mx.symbol.SequenceReverse}
\title{SequenceReverse:Reverses the elements of each sequence. Takes an n-dimensional tensor of the form [max sequence length, batchsize, other dims] and returns a tensor of the same shape. This operator takes an optional input tensor sequence_length of positive ints of dimension [batchsize] when the sequence_length option is set to true. This allows the operator to handle variable-length sequences. If sequence_length is false, then each example in the batch is assumed to have the max sequence length.}
\usage{
mx.symbol.SequenceReverse(...)
}
\arguments{
\item{data}{Symbol
n-dimensional input tensor of the form [max sequence length, batchsize, other dims]}

\item{sequence.length}{Symbol
vector of sequence lengths of size batchsize}

\item{use.sequence.length}{boolean, optional, default=False
If set to true, this layer takes in extra input sequence_length to specify variable length sequence}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
SequenceReverse:Reverses the elements of each sequence. Takes an n-dimensional tensor of the form [max sequence length, batchsize, other dims] and returns a tensor of the same shape. This operator takes an optional input tensor sequence_length of positive ints of dimension [batchsize] when the sequence_length option is set to true. This allows the operator to handle variable-length sequences. If sequence_length is false, then each example in the batch is assumed to have the max sequence length.
}

