% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.sgd_mom_update}
\alias{mx.symbol.sgd_mom_update}
\title{sgd_mom_update:Updater function for sgd optimizer}
\usage{
mx.symbol.sgd_mom_update(...)
}
\arguments{
\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
sgd_mom_update:Updater function for sgd optimizer
}

