% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SliceChannel}
\alias{mx.symbol.SliceChannel}
\title{Slice channel into many outputs with equally divided channel}
\usage{
mx.symbol.SliceChannel(...)
}
\arguments{
\item{num.outputs}{int, required
Number of outputs to be sliced.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Slice channel into many outputs with equally divided channel
}

