% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.smooth_l1}
\alias{mx.symbol.smooth_l1}
\title{smooth_l1:Calculate Smooth L1 Loss(lhs, scalar)}
\usage{
mx.symbol.smooth_l1(...)
}
\arguments{
\item{lhs}{NDArray
source input}

\item{scalar}{float
scalar input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/elemwise_binary_scalar_op.cc:98
}

