% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Softmax}
\alias{mx.symbol.Softmax}
\title{Perform a softmax transformation on input.}
\usage{
mx.symbol.Softmax(...)
}
\arguments{
\item{data}{Symbol
Input data to softmax.}

\item{grad.scale}{float, optional, default=1
Scale the gradient by a float factor}

\item{multi.output}{boolean, optional, default=False
If set to true, for a (n,k,x_1,..,x_n) dimensionalinput tensor, softmax will generate n*x_1*...*x_n output, eachhas k classes}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Perform a softmax transformation on input.
}

