% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.softmax_cross_entropy}
\alias{mx.symbol.softmax_cross_entropy}
\title{Calculate cross_entropy(lhs, one_hot(rhs))}
\usage{
mx.symbol.softmax_cross_entropy(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Right symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Calculate cross_entropy(lhs, one_hot(rhs))
}

