% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SpatialTransformer}
\alias{mx.symbol.SpatialTransformer}
\title{Apply spatial transformer to input feature map.}
\usage{
mx.symbol.SpatialTransformer(...)
}
\arguments{
\item{data}{Symbol
Input data to the SpatialTransformerOp.}

\item{loc}{Symbol
localisation net, the output dim should be 6 when transform_type is affine, and the name of loc symbol should better starts with 'stn_loc', so that initialization it with iddentify tranform, or you shold initialize the weight and bias by yourself.}

\item{target.shape}{Shape(tuple), optional, default=(0,0)
output shape(h, w) of spatial transformer: (y, x)}

\item{transform.type}{{'affine'}, required
transformation type}

\item{sampler.type}{{'bilinear'}, required
sampling type}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply spatial transformer to input feature map.
}

