% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.square}
\alias{mx.symbol.square}
\title{Take square of the src}
\usage{
mx.symbol.square(...)
}
\arguments{
\item{src}{Symbol
Source symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Take square of the src
}

