% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.sum}
\alias{mx.symbol.sum}
\title{Take sum of the src.The result will be ndarray of shape (1,) on the same device.}
\usage{
mx.symbol.sum(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Left symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Take sum of the src.The result will be ndarray of shape (1,) on the same device.
}

