% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.sum_axis}
\alias{mx.symbol.sum_axis}
\title{(Depreciated! Use sum instead!) Take sum of the src in the given axis and returns a NDArray. Follows numpy semantics.}
\usage{
mx.symbol.sum_axis(...)
}
\arguments{
\item{src}{Symbol
Left symbolic input to the function}

\item{axis}{Shape(tuple), optional, default=()
Same as Numpy. The axes to perform the reduction.If left empty, a global reduction will be performed.}

\item{keepdims}{boolean, optional, default=False
Same as Numpy. If keepdims is set to true, the axis which is reduced is left in the result as dimension with size one.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
(Depreciated! Use sum instead!) Take sum of the src in the given axis and returns a NDArray. Follows numpy semantics.
}

