% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SVMOutput}
\alias{mx.symbol.SVMOutput}
\title{SVMOutput:Support Vector Machine based transformation on input, backprop L2-SVM}
\usage{
mx.symbol.SVMOutput(...)
}
\arguments{
\item{data}{Symbol
Input data to svm.}

\item{label}{Symbol
Label data.}

\item{margin}{float, optional, default=1
Scale the DType(param_.margin) for activation size}

\item{regularization.coefficient}{float, optional, default=1
Scale the coefficient responsible for balacing coefficient size and error tradeoff}

\item{use.linear}{boolean, optional, default=False
If set true, uses L1-SVM objective function. Default uses L2-SVM objective}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
SVMOutput:Support Vector Machine based transformation on input, backprop L2-SVM
}

