% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SwapAxis}
\alias{mx.symbol.SwapAxis}
\title{Apply swapaxis to input.}
\usage{
mx.symbol.SwapAxis(...)
}
\arguments{
\item{data}{Symbol
Input data to the SwapAxisOp.}

\item{dim1}{int (non-negative), optional, default=0
the first axis to be swapped.}

\item{dim2}{int (non-negative), optional, default=0
the second axis to be swapped.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply swapaxis to input.
}

