% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.transpose}
\alias{mx.symbol.transpose}
\title{Transpose the input matrix and return a new one}
\usage{
mx.symbol.transpose(...)
}
\arguments{
\item{lhs}{Symbol
Left symbolic input to the function}

\item{rhs}{Symbol
Left symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Transpose the input matrix and return a new one
}

