% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.uniform}
\alias{mx.symbol.uniform}
\title{Sample a uniform distribution}
\usage{
mx.symbol.uniform(...)
}
\arguments{
\item{low}{float, optional, default=0
The lower bound of distribution}

\item{high}{float, optional, default=1
The upper bound of distribution}

\item{shape}{Shape(tuple), required
The shape of the output}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Sample a uniform distribution
}

