% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mx.symbol.Variable}
\alias{mx.symbol.Variable}
\title{Create a symbolic variable with specified name.}
\arguments{
\item{name}{string
The name of the result symbol.}
}
\value{
The result symbol
}
\description{
Create a symbolic variable with specified name.
}

