\name{mxnet-package}
\alias{mxnet-package}
\alias{mxnet}
\docType{package}
\title{
 MXNet is a deep learning framework designed for both efficiency and flexibility. 
}
\description{
  MXNet is a deep learning framework designed for both efficiency and flexibility. It allows you to mix the flavours of deep learning programs together to maximize the efficiency and your productivity.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Tianqi Chen, Qiang Kou

Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{ package }
