% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mxnet.export}
\alias{mxnet.export}
\title{Internal function to generate mxnet_generated.R
Users do not need to call this function.}
\usage{
mxnet.export(path)
}
\arguments{
\item{path}{The path to the root of the package.}
}
\description{
Internal function to generate mxnet_generated.R
Users do not need to call this function.
}

