% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mxnet}
\alias{mxnet}
\alias{mxnet-package}
\title{MXNet: Flexible and Efficient GPU computing and Deep Learning.}
\description{
MXNet is a flexible and efficient GPU computing and deep learning framework.
}
\details{
It enables you to write seamless tensor/matrix computation with multiple GPUs in R.

It also enables you construct and customize the state-of-art deep learning models in R,
and apply them to tasks such as image classification and data science challenges.
}

