% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{Ops.MXNDArray}
\alias{Ops.MXNDArray}
\title{Binary operator overloading of mx.ndarray}
\usage{
\method{Ops}{MXNDArray}(e1, e2)
}
\arguments{
\item{e1}{The first operand}

\item{e1}{The second operand}
}
\description{
Binary operator overloading of mx.ndarray
}

