% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{predict.MXFeedForwardModel}
\alias{predict.MXFeedForwardModel}
\title{Predict the outputs given a model and dataset.}
\usage{
\method{predict}{MXFeedForwardModel}(model, X, ctx = NULL,
  array.batch.size = 128)
}
\arguments{
\item{model}{The MXNet Model.}

\item{X}{The dataset to predict.}

\item{ctx}{mx.cpu() or mx.gpu(i) The device used to generate the prediction.}

\item{array.batch.size}{The batch size used in batching. Only used when X is R's array.}
}
\description{
Predict the outputs given a model and dataset.
}

