% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndarray.R
\name{print.MXNDArray}
\alias{print.MXNDArray}
\title{print operator overload of mx.ndarray}
\usage{
\method{print}{MXNDArray}(nd)
}
\arguments{
\item{nd}{The mx.ndarray}
}
\description{
print operator overload of mx.ndarray
}

