% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{mx.callback.log.train.metric}
\alias{mx.callback.log.train.metric}
\title{Log training metric each period}
\usage{
mx.callback.log.train.metric(period, logger = NULL)
}
\arguments{
\item{period}{The number of batch to log the training evaluation metric}

\item{logger}{The logger class}
}
\description{
Log training metric each period
}

