% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.mae}
\alias{mx.metric.mae}
\title{MAE (Mean Absolute Error) metric for regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.mae
}
\description{
MAE (Mean Absolute Error) metric for regression
}
\keyword{datasets}

