% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.rmse}
\alias{mx.metric.rmse}
\title{RMSE (Root Mean Squared Error) metric for regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.rmse
}
\description{
RMSE (Root Mean Squared Error) metric for regression
}
\keyword{datasets}

