% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.rmsle}
\alias{mx.metric.rmsle}
\title{RMSLE (Root Mean Squared Logarithmic Error) metric for regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.rmsle
}
\description{
RMSLE (Root Mean Squared Logarithmic Error) metric for regression
}
\keyword{datasets}

