% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.batch.dot}
\alias{mx.nd.batch.dot}
\title{Calculate batched dot product of two matrices. (batch, M, K) batch_dot (batch, K, N) --> (batch, M, N)}
\arguments{
\item{lhs}{NDArray
Left input}

\item{rhs}{NDArray
Right input}

\item{axis}{int, required
The dimension to flip}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/matrix_op.cc:176
}

