% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.axis}
\alias{mx.nd.broadcast.axis}
\title{Broadcast data in the given axis to the given size. The original size of the broadcasting axis must be 1.}
\arguments{
\item{src}{NDArray
Source input to the function}

\item{axis}{Shape(tuple), optional, default=()
The axes to perform the broadcasting.}

\item{size}{Shape(tuple), optional, default=()
Target sizes of the broadcasting axes.}
}
\value{
out The result mx.ndarray
}
\description{
Broadcast data in the given axis to the given size. The original size of the broadcasting axis must be 1.
}

