% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.ceil}
\alias{mx.nd.ceil}
\title{Take ceil value of the src}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Take ceil value of the src
}

