% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.crop}
\alias{mx.nd.crop}
\title{Crop the input matrix and return a new one}
\arguments{
\item{src}{NDArray
Source input to the function}

\item{begin}{Shape(tuple), required
starting coordinates}

\item{end}{Shape(tuple), required
ending coordinates}
}
\value{
out The result mx.ndarray
}
\description{
Crop the input matrix and return a new one
}

