% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.dot}
\alias{mx.nd.dot}
\title{Calculate 2D matrix multiplication}
\arguments{
\item{lhs}{NDArray
Left operand to the function.}

\item{rhs}{NDArray
Right operand to the function.}
}
\value{
out The result mx.ndarray
}
\description{
Calculate 2D matrix multiplication
}

