% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.expand.dims}
\alias{mx.nd.expand.dims}
\title{Expand the shape of array by inserting a new axis.}
\arguments{
\item{src}{NDArray
Source input to the function}

\item{axis}{int (non-negative), required
Position (amongst axes) where new axis is to be inserted.}
}
\value{
out The result mx.ndarray
}
\description{
Expand the shape of array by inserting a new axis.
}

