% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.flip}
\alias{mx.nd.flip}
\title{Flip the input tensor along axis and return a new one.}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int, required
The dimension to flip}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/matrix_op.cc:142
}

