% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.smooth.l1}
\alias{mx.nd.smooth.l1}
\title{Calculate Smooth L1 Loss(lhs, scalar)}
\arguments{
\item{src}{NDArray
Source input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Calculate Smooth L1 Loss(lhs, scalar)
}

