% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.softmax.cross.entropy}
\alias{mx.nd.softmax.cross.entropy}
\title{Calculate cross_entropy(lhs, one_hot(rhs))}
\arguments{
\item{lhs}{NDArray
Left operand  to the function}

\item{rhs}{NDArray
Right operand to the function}
}
\value{
out The result mx.ndarray
}
\description{
Calculate cross_entropy(lhs, one_hot(rhs))
}

