% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sum}
\alias{mx.nd.sum}
\title{Take sum of the src in the given axis and returns a NDArray. Follows numpy semantics.}
\arguments{
\item{src}{NDArray
Source input to the function}

\item{axis}{Shape(tuple), optional, default=()
Same as Numpy. The axes to perform the reduction.If left empty, a global reduction will be performed.}

\item{keepdims}{boolean, optional, default=False
Same as Numpy. If keepdims is set to true, the axis which is reduced is left in the result as dimension with size one.}
}
\value{
out The result mx.ndarray
}
\description{
Take sum of the src in the given axis and returns a NDArray. Follows numpy semantics.
}

