% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.transpose}
\alias{mx.nd.transpose}
\title{Transpose the input tensor and return a new one}
\arguments{
\item{src}{NDArray
Source input}

\item{axes}{Shape(tuple), optional, default=()
Target axis order. By default the axes will be inverted.}
}
\value{
out The result mx.ndarray
}
\description{
From:src/operator/tensor/matrix_op.cc:20
}

