% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Activation}
\alias{mx.symbol.Activation}
\title{Apply activation function to input.}
\usage{
mx.symbol.Activation(...)
}
\arguments{
\item{data}{Symbol
Input data to activation function.}

\item{act.type}{{'relu', 'sigmoid', 'tanh'}, required
Activation function to be applied.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply activation function to input.
}

