% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.batch_dot}
\alias{mx.symbol.batch_dot}
\title{batch_dot:Calculate batched dot product of two matrices. (batch, M, K) batch_dot (batch, K, N) --> (batch, M, N)}
\usage{
mx.symbol.batch_dot(...)
}
\arguments{
\item{lhs}{NDArray
Left input}

\item{rhs}{NDArray
Right input}

\item{axis}{int, required
The dimension to flip}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/matrix_op.cc:176
}

