% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.BatchNorm}
\alias{mx.symbol.BatchNorm}
\title{Apply batch normalization to input.}
\usage{
mx.symbol.BatchNorm(...)
}
\arguments{
\item{data}{Symbol
Input data to batch normalization}

\item{eps}{float, optional, default=0.001
Epsilon to prevent div 0}

\item{momentum}{float, optional, default=0.9
Momentum for moving average}

\item{fix.gamma}{boolean, optional, default=True
Fix gamma while training}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply batch normalization to input.
}

