% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_axis}
\alias{mx.symbol.broadcast_axis}
\title{broadcast_axis:Broadcast src along axis}
\usage{
mx.symbol.broadcast_axis(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{Shape(tuple), optional, default=()
The axes to perform the broadcasting.}

\item{size}{Shape(tuple), optional, default=()
Target sizes of the broadcasting axes.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/broadcast_reduce_op.cc:43
}

