% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_power}
\alias{mx.symbol.broadcast_power}
\title{broadcast_power:}
\usage{
mx.symbol.broadcast_power(...)
}
\arguments{
\item{lhs}{NDArray
first input}

\item{rhs}{NDArray
second input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
broadcast_power:
}

