% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_to}
\alias{mx.symbol.broadcast_to}
\title{Broadcast data to the target shape. The original size of the broadcasting axis must be 1.}
\usage{
mx.symbol.broadcast_to(...)
}
\arguments{
\item{src}{Symbol
Left symbolic input to the function}

\item{shape}{Shape(tuple), optional, default=()
The shape of the desired array. We can set the dim to zero if it's same as the original. E.g `A = broadcast_to(B, shape=(10, 0, 0))` has the same meaning as `A = broadcast_axis(B, axis=0, size=10)`.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Broadcast data to the target shape. The original size of the broadcasting axis must be 1.
}

