% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Convolution}
\alias{mx.symbol.Convolution}
\title{Apply convolution to input then add a bias.}
\usage{
mx.symbol.Convolution(...)
}
\arguments{
\item{data}{Symbol
Input data to the ConvolutionOp.}

\item{weight}{Symbol
Weight matrix.}

\item{bias}{Symbol
Bias parameter.}

\item{kernel}{Shape(tuple), required
convolution kernel size: (y, x)}

\item{stride}{Shape(tuple), optional, default=(1,1)
convolution stride: (y, x)}

\item{dilate}{Shape(tuple), optional, default=(1,1)
convolution dilate: (y, x)}

\item{pad}{Shape(tuple), optional, default=(0,0)
pad for convolution: (y, x)}

\item{num.filter}{int (non-negative), required
convolution filter(channel) number}

\item{num.group}{int (non-negative), optional, default=1
Number of groups partition. This option is not supported by CuDNN, you can use SliceChannel to num_group,apply convolution and concat instead to achieve the same need.}

\item{workspace}{long (non-negative), optional, default=512
Tmp workspace for convolution (MB).}

\item{no.bias}{boolean, optional, default=False
Whether to disable bias parameter.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply convolution to input then add a bias.
}

