% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Crop}
\alias{mx.symbol.Crop}
\title{Crop the 2nd and 3rd dim of input data, with the corresponding size of w_h or with width and height of the second input symbol}
\usage{
mx.symbol.Crop(...)
}
\arguments{
\item{num.args}{int, required
Number of inputs for crop, if equals one, then we will use the h_wfor crop heihgt and width, else if equals two, then we will use the heightand width of the second input symbol, we name crop_like here}

\item{offset}{Shape(tuple), optional, default=(0,0)
crop offset coordinate: (y, x)}

\item{h.w}{Shape(tuple), optional, default=(0,0)
crop height and weight: (h, w)}

\item{center.crop}{boolean, optional, default=False
If set to true, then it will use be the center_crop,or it will crop using the shape of crop_like}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Crop the 2nd and 3rd dim of input data, with the corresponding size of w_h or with width and height of the second input symbol
}

