% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.dot}
\alias{mx.symbol.dot}
\title{dot:Calculate dot product of two matrices or two vectors.}
\usage{
mx.symbol.dot(...)
}
\arguments{
\item{lhs}{NDArray
Left input}

\item{rhs}{NDArray
Right input}

\item{axis}{int, required
The dimension to flip}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/matrix_op.cc:154
}

