% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Dropout}
\alias{mx.symbol.Dropout}
\title{Apply dropout to input}
\usage{
mx.symbol.Dropout(...)
}
\arguments{
\item{data}{Symbol
Input data to dropout.}

\item{p}{float, optional, default=0.5
Fraction of the input that gets dropped out at training time}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Apply dropout to input
}

