% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.ElementWiseSum}
\alias{mx.symbol.ElementWiseSum}
\title{Perform an elementwise sum over all the inputs.}
\usage{
mx.symbol.ElementWiseSum(...)
}
\arguments{
\item{num.args}{int, required
Number of inputs to be summed.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Perform an elementwise sum over all the inputs.
}

