% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Embedding}
\alias{mx.symbol.Embedding}
\title{Get embedding for one-hot input. A n-dimensional input tensor will be trainsformed into a (n+1)-dimensional tensor, where a new dimension is added for the embedding results.}
\usage{
mx.symbol.Embedding(...)
}
\arguments{
\item{data}{Symbol
Input data to the EmbeddingOp.}

\item{weight}{Symbol
Enbedding weight matrix.}

\item{input.dim}{int, required
input dim of one-hot encoding}

\item{output.dim}{int, required
output dim of embedding}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Get embedding for one-hot input. A n-dimensional input tensor will be trainsformed into a (n+1)-dimensional tensor, where a new dimension is added for the embedding results.
}

