% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.expand_dims}
\alias{mx.symbol.expand_dims}
\title{expand_dims:Expand the shape of array by inserting a new axis.}
\usage{
mx.symbol.expand_dims(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{axis}{int (non-negative), required
Position (amongst axes) where new axis is to be inserted.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/matrix_op.cc:48
}

