% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Flatten}
\alias{mx.symbol.Flatten}
\title{Flatten input}
\usage{
mx.symbol.Flatten(...)
}
\arguments{
\item{data}{Symbol
Input data to  flatten.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Flatten input
}

