% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.L2Normalization}
\alias{mx.symbol.L2Normalization}
\title{L2Normalization:Set the l2 norm of each instance to a constant.}
\usage{
mx.symbol.L2Normalization(...)
}
\arguments{
\item{data}{Symbol
Input data to the L2NormalizationOp.}

\item{eps}{float, optional, default=1e-10
Epsilon to prevent div 0}

\item{mode}{{'channel', 'instance', 'spatial'},optional, default='instance'
Normalization Mode. If set to instance, this operator will compute a norm for each instance in the batch; this is the default mode. If set to channel, this operator will compute a cross channel norm at each position of each instance. If set to spatial, this operator will compute a norm for each channel.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
L2Normalization:Set the l2 norm of each instance to a constant.
}

