% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.MakeLoss}
\alias{mx.symbol.MakeLoss}
\title{MakeLoss:Get output from a symbol and pass 1 gradient back. This is used as a terminal loss if unary and binary operator are used to composite a loss with no declaration of backward dependency}
\usage{
mx.symbol.MakeLoss(...)
}
\arguments{
\item{data}{Symbol
Input data.}

\item{grad.scale}{float, optional, default=1
gradient scale as a supplement to unary and binary operators}

\item{valid.thresh}{float, optional, default=0
regard element valid when x > valid_thresh, this is used only in valid normalization mode.}

\item{normalization}{{'batch', 'null', 'valid'},optional, default='null'
If set to null, op will not normalize on output gradient.If set to batch, op will normalize gradient by divide batch size.If set to valid, op will normalize gradient by divide # sample marked as valid}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
MakeLoss:Get output from a symbol and pass 1 gradient back. This is used as a terminal loss if unary and binary operator are used to composite a loss with no declaration of backward dependency
}

