% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.normal}
\alias{mx.symbol.normal}
\title{Sample a normal distribution}
\usage{
mx.symbol.normal(...)
}
\arguments{
\item{loc}{float, optional, default=0
Mean of the distribution.}

\item{scale}{float, optional, default=1
Standard deviation of the distribution.}

\item{shape}{Shape(tuple), required
The shape of the output}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Sample a normal distribution
}

