% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Pooling}
\alias{mx.symbol.Pooling}
\title{Perform spatial pooling on inputs.}
\usage{
mx.symbol.Pooling(...)
}
\arguments{
\item{data}{Symbol
Input data to the pooling operator.}

\item{kernel}{Shape(tuple), required
pooling kernel size: (y, x)}

\item{pool.type}{{'avg', 'max', 'sum'}, required
Pooling type to be applied.}

\item{stride}{Shape(tuple), optional, default=(1,1)
stride: for pooling (y, x)}

\item{pad}{Shape(tuple), optional, default=(0,0)
pad for pooling: (y, x)}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Perform spatial pooling on inputs.
}

