% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Reshape}
\alias{mx.symbol.Reshape}
\title{Reshape input to target shape}
\usage{
mx.symbol.Reshape(...)
}
\arguments{
\item{data}{Symbol
Input data to  reshape.}

\item{target.shape}{Shape(tuple), required
Target new shape. One and only one dim can be 0, in which case it will be inferred from the rest of dims}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Reshape input to target shape
}

