% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.smooth_l1}
\alias{mx.symbol.smooth_l1}
\title{Calculate Smooth L1 Loss(lhs, scalar)}
\usage{
mx.symbol.smooth_l1(...)
}
\arguments{
\item{src}{Symbol
Left symbolic input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Calculate Smooth L1 Loss(lhs, scalar)
}

