% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.Softmax}
\alias{mx.symbol.Softmax}
\title{DEPRECATED: Perform a softmax transformation on input. Please use SoftmaxOutput}
\usage{
mx.symbol.Softmax(...)
}
\arguments{
\item{data}{Symbol
Input data to softmax.}

\item{grad.scale}{float, optional, default=1
Scale the gradient by a float factor}

\item{ignore.label}{float, optional, default=-1
the label value will be ignored during backward (only works if use_ignore is set to be true).}

\item{multi.output}{boolean, optional, default=False
If set to true, for a (n,k,x_1,..,x_n) dimensional input tensor, softmax will generate n*x_1*...*x_n output, each has k classes}

\item{use.ignore}{boolean, optional, default=False
If set to true, the ignore_label value will not contribute to the backward gradient}

\item{normalization}{{'batch', 'null', 'valid'},optional, default='null'
If set to null, op will do nothing on output gradient.If set to batch, op will normalize gradient by divide batch sizeIf set to valid, op will normalize gradient by divide sample not ignored}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
DEPRECATED: Perform a softmax transformation on input. Please use SoftmaxOutput
}

