% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.softmax_cross_entropy}
\alias{mx.symbol.softmax_cross_entropy}
\title{softmax_cross_entropy:Calculate cross_entropy(lhs, one_hot(rhs))}
\usage{
mx.symbol.softmax_cross_entropy(...)
}
\arguments{
\item{data}{NDArray
Input data}

\item{label}{NDArray
Input label}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/loss_binary_op.cc:12
}

