% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.transpose}
\alias{mx.symbol.transpose}
\title{transpose:Transpose the input tensor and return a new one}
\usage{
mx.symbol.transpose(...)
}
\arguments{
\item{src}{NDArray
Source input}

\item{axes}{Shape(tuple), optional, default=()
Target axis order. By default the axes will be inverted.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
From:src/operator/tensor/matrix_op.cc:20
}

